neural-network-stock-predictor
Semester project for COS 470 (Artifical Intelligence) at the University
of Maine.

This is an application that utilizes the power of back-propagating
neural networks in order to attempt to predict future stock prices of a
specified stock symbol. Although the power and accuracy of neural
networks are based upon the input and training you provide it, they
would not usually be the best option for attempting to predict stock
prices as we're doing here.